+++ docs/scripts/views/autocomplete.vue
@@ -49,12 +49,13 @@ export default {
       this.suggestion = data[this.text] ? data[this.text].map((item, index) => {
         return {
           active: index === 0,
+          key: index,
           value: item
         };
       }) : [];
     },
-    onInputEnter(val) {
-      this.text = val;
+    onInputEnter(data) {
+      this.text = data.value;
     }
   },
   created() {